extract leave room button
